mimalloc: allow running in Windows Nano Server containers
The `GetNumaProcessorNode()` symbol is not defined in Nano Server's DLLs
(because that function is long deprecated). This caused Git no longer to
work on Nano Server.

Instead of importing it statically, try to import it dynamically, and
fall back gracefully if it cannot be loaded.

This fixes https://github.com/git-for-windows/git/issues/4052

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>